# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tenant/{tenantId}/lead-lists/{listId}/campaigns": {
      "get": {
        "description": "Inclui relação mínima dos templates (id, name, language, status) e bindings persistidos.",
        "operationId": "ListCampaignsController_listCampaigns",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "ID numérico do tenant",
            "schema": {
              "example": 8,
              "type": "number"
            }
          },
          {
            "name": "listId",
            "required": true,
            "in": "path",
            "description": "ID da lista de leads do tenant",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListCampaignResponseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tenant ou lista não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar campanhas da lista",
        "tags": [
          "Tenant — List Campaigns"
        ]
      },
      "post": {
        "description": "enabled=true exige template APPROVED, bindings `send` completos, notifyTemplate + bindings `notify` se algum buttonAction for NOTIFY, e labels QUICK_REPLY válidos no template de disparo.",
        "operationId": "ListCampaignsController_createCampaign",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "ID numérico do tenant",
            "schema": {
              "example": 8,
              "type": "number"
            }
          },
          {
            "name": "listId",
            "required": true,
            "in": "path",
            "description": "ID da lista de leads do tenant",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertListCampaignDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCampaignResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validação de enable, bindings ou buttonActions"
          },
          "404": {
            "description": "Tenant ou lista não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar campanha",
        "tags": [
          "Tenant — List Campaigns"
        ]
      }
    },
    "/tenant/{tenantId}/lead-lists/{listId}/campaigns/{campaignId}": {
      "get": {
        "operationId": "ListCampaignsController_getCampaign",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "ID numérico do tenant",
            "schema": {
              "example": 8,
              "type": "number"
            }
          },
          {
            "name": "listId",
            "required": true,
            "in": "path",
            "description": "ID da lista de leads do tenant",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "description": "ID da campanha na lista",
            "schema": {
              "example": 3,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCampaignResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant, lista ou campanha não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter campanha por id",
        "tags": [
          "Tenant — List Campaigns"
        ]
      },
      "put": {
        "description": "Mesmas regras de validação do POST.",
        "operationId": "ListCampaignsController_replaceCampaign",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "ID numérico do tenant",
            "schema": {
              "example": 8,
              "type": "number"
            }
          },
          {
            "name": "listId",
            "required": true,
            "in": "path",
            "description": "ID da lista de leads do tenant",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "description": "ID da campanha na lista",
            "schema": {
              "example": 3,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertListCampaignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCampaignResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validação falhou"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Substituir campanha (corpo completo)",
        "tags": [
          "Tenant — List Campaigns"
        ]
      },
      "patch": {
        "description": "Merge com valores existentes. enabled=true revalida templates e bindings resultantes.",
        "operationId": "ListCampaignsController_patchCampaign",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "ID numérico do tenant",
            "schema": {
              "example": 8,
              "type": "number"
            }
          },
          {
            "name": "listId",
            "required": true,
            "in": "path",
            "description": "ID da lista de leads do tenant",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "description": "ID da campanha na lista",
            "schema": {
              "example": 3,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchListCampaignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCampaignResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validação falhou"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar campanha (parcial)",
        "tags": [
          "Tenant — List Campaigns"
        ]
      }
    },
    "/tenant/{tenantId}/lead-lists/{listId}/sends": {
      "get": {
        "description": "Até 100 registros, ordenados por sentAt desc. status=failed filtra por lastStatus failed e inclui latestError quando disponível.",
        "operationId": "ListSendsController_listSends",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "ID numérico do tenant",
            "schema": {
              "example": 8,
              "type": "number"
            }
          },
          {
            "name": "listId",
            "required": true,
            "in": "path",
            "description": "ID da lista de leads do tenant",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filtrar apenas sends com falha de entrega Meta",
            "schema": {
              "enum": [
                "failed"
              ],
              "type": "string"
            }
          },
          {
            "name": "campaignId",
            "required": false,
            "in": "query",
            "description": "Filtrar sends de uma campanha específica",
            "schema": {
              "example": 3,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListSendResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "campaignId inválido"
          },
          "404": {
            "description": "Tenant ou lista não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar envios (sends) da lista",
        "tags": [
          "Tenant — List Campaigns"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gestão de Leads",
    "description": "API para gestão de leads",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CatalogTemplateRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 5
          },
          "name": {
            "type": "string",
            "example": "outreach_construtoras"
          },
          "language": {
            "type": "string",
            "example": "pt_BR"
          },
          "status": {
            "type": "string",
            "example": "APPROVED"
          }
        },
        "required": [
          "id",
          "name",
          "language",
          "status"
        ]
      },
      "ListCampaignResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 3
          },
          "listId": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Campanha A"
          },
          "enabled": {
            "type": "boolean",
            "example": false
          },
          "templateId": {
            "type": "number",
            "example": 5
          },
          "slotBindings": {
            "type": "object",
            "description": "Bindings do template de disparo (role send)",
            "example": {
              "send": {
                "body.1": {
                  "type": "literal",
                  "value": "Olá"
                },
                "body.customer_name": {
                  "type": "recipient.name"
                }
              }
            }
          },
          "notifyTemplateId": {
            "type": "number",
            "nullable": true,
            "example": 6
          },
          "notifySlotBindings": {
            "type": "object",
            "description": "Bindings do template de notify (role notify)",
            "example": {
              "notify": {
                "body.customer_name": {
                  "type": "recipient.name"
                },
                "body.customer_phone": {
                  "type": "recipient.phone"
                }
              }
            }
          },
          "buttonActions": {
            "type": "array",
            "example": [
              {
                "buttonIndex": 0,
                "label": "Tenho Interesse!",
                "action": "NOTIFY"
              },
              {
                "buttonIndex": 1,
                "label": "Agora não",
                "action": "NOOP"
              }
            ]
          },
          "schedule": {
            "type": "object",
            "description": "Mapa dia-da-semana (0=Dom … 6=Sáb) → horas UTC",
            "example": {
              "2": [
                18
              ],
              "4": [
                13
              ]
            }
          },
          "sendsPerRun": {
            "type": "number",
            "example": 5
          },
          "sendIntervalSeconds": {
            "type": "number",
            "example": 5
          },
          "template": {
            "$ref": "#/components/schemas/CatalogTemplateRefDto"
          },
          "notifyTemplate": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/CatalogTemplateRefDto"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-08-17T20:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-08-17T20:00:00.000Z"
          }
        },
        "required": [
          "id",
          "listId",
          "name",
          "enabled",
          "templateId",
          "slotBindings",
          "notifySlotBindings",
          "buttonActions",
          "schedule",
          "sendsPerRun",
          "sendIntervalSeconds",
          "template",
          "createdAt",
          "updatedAt"
        ]
      },
      "SendSlotBindingsDto": {
        "type": "object",
        "properties": {
          "send": {
            "type": "object",
            "description": "Bindings do template de disparo, keyed por slot",
            "example": {
              "body.1": {
                "type": "literal",
                "value": "Olá"
              },
              "body.customer_name": {
                "type": "recipient.name"
              }
            }
          }
        },
        "required": [
          "send"
        ]
      },
      "NotifySlotBindingsDto": {
        "type": "object",
        "properties": {
          "notify": {
            "type": "object",
            "description": "Bindings do template de notify, keyed por slot",
            "example": {
              "body.customer_name": {
                "type": "recipient.name"
              }
            }
          }
        },
        "required": [
          "notify"
        ]
      },
      "ButtonActionDto": {
        "type": "object",
        "properties": {
          "buttonIndex": {
            "type": "number",
            "minimum": 0,
            "example": 0
          },
          "label": {
            "type": "string",
            "example": "Tenho Interesse!"
          },
          "action": {
            "type": "string",
            "enum": [
              "NOTIFY",
              "NOOP"
            ],
            "example": "NOTIFY"
          }
        },
        "required": [
          "buttonIndex",
          "label",
          "action"
        ]
      },
      "UpsertListCampaignDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Campanha A"
          },
          "enabled": {
            "type": "boolean",
            "description": "Habilitar execução no cron (validação estrita quando true)",
            "default": false,
            "example": false
          },
          "templateId": {
            "type": "number",
            "minimum": 1,
            "description": "FK do template de disparo no catálogo"
          },
          "slotBindings": {
            "$ref": "#/components/schemas/SendSlotBindingsDto"
          },
          "notifyTemplateId": {
            "type": "number",
            "minimum": 1,
            "description": "FK do template de notify (obrigatório se NOTIFY)"
          },
          "notifySlotBindings": {
            "$ref": "#/components/schemas/NotifySlotBindingsDto"
          },
          "buttonActions": {
            "minItems": 0,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ButtonActionDto"
            }
          },
          "schedule": {
            "type": "object",
            "description": "Mapa dia-da-semana → horas UTC",
            "example": {
              "2": [
                18
              ],
              "4": [
                13
              ]
            }
          },
          "sendsPerRun": {
            "type": "number",
            "minimum": 1,
            "default": 5
          },
          "sendIntervalSeconds": {
            "type": "number",
            "minimum": 0,
            "default": 5
          }
        },
        "required": [
          "name",
          "templateId",
          "slotBindings",
          "buttonActions",
          "schedule"
        ]
      },
      "PatchListCampaignDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Campanha A"
          },
          "enabled": {
            "type": "boolean"
          },
          "templateId": {
            "type": "number",
            "minimum": 1
          },
          "slotBindings": {
            "$ref": "#/components/schemas/SendSlotBindingsDto"
          },
          "notifyTemplateId": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "notifySlotBindings": {
            "$ref": "#/components/schemas/NotifySlotBindingsDto"
          },
          "buttonActions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ButtonActionDto"
            }
          },
          "schedule": {
            "type": "object",
            "description": "Mapa dia-da-semana → horas UTC",
            "example": {
              "2": [
                18
              ],
              "4": [
                13
              ]
            }
          },
          "sendsPerRun": {
            "type": "number",
            "minimum": 1
          },
          "sendIntervalSeconds": {
            "type": "number",
            "minimum": 0
          }
        }
      },
      "ListSendLeadRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 42
          },
          "name": {
            "type": "string",
            "example": "Construtora ABC"
          },
          "phone": {
            "type": "string",
            "example": "5511987654321"
          }
        },
        "required": [
          "id",
          "name",
          "phone"
        ]
      },
      "ListSendCampaignRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 3
          },
          "name": {
            "type": "string",
            "example": "Campanha A"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "ListSendResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 100
          },
          "wamid": {
            "type": "string",
            "description": "ID da mensagem outbound na Meta (wamid)",
            "example": "wamid.HBgMNTUxMTk4NzY1NDMyMRUCABIYFDN..."
          },
          "sentAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-08-17T20:05:00.000Z"
          },
          "lastStatus": {
            "type": "string",
            "nullable": true,
            "enum": [
              "sent",
              "delivered",
              "read",
              "failed"
            ],
            "example": "delivered"
          },
          "listLead": {
            "$ref": "#/components/schemas/ListSendLeadRefDto"
          },
          "campaign": {
            "$ref": "#/components/schemas/ListSendCampaignRefDto"
          },
          "latestError": {
            "type": "object",
            "description": "Presente quando lastStatus=failed",
            "example": {
              "code": 131026,
              "title": "Message undeliverable"
            }
          }
        },
        "required": [
          "id",
          "wamid",
          "sentAt",
          "listLead",
          "campaign"
        ]
      },
...
    }
  }
}